# docs/swagger.yaml
basePath: /
definitions:
  bookmark.createBookmarkRequest:
    properties:
      description:
        example: A sample bookmark
        type: string
      url:
        example: https://example.com
        type: string
    required:
    - description
    - url
    type: object
  bookmark.createBookmarkResponse:
    properties:
      data:
        $ref: '#/definitions/model.Bookmark'
      message:
        type: string
    type: object
  bookmark.updateBookmarkRequest:
    properties:
      description:
        type: string
      url:
        type: string
    required:
    - description
    - url
    type: object
  common.Message:
    properties:
      details: {}
      message:
        type: string
    type: object
  common.Paging:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  healthcheck.healthCheckResponse:
    properties:
      instance_id:
        type: string
      message:
        type: string
      service_name:
        type: string
    type: object
  link.shortenURLRequest:
    properties:
      exp:
        maximum: 604800
        type: integer
      url:
        type: string
    required:
    - exp
    - url
    type: object
  link.shortenURLResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  model.Bookmark:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
  model.User:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  user.createUserRequest:
    properties:
      display_name:
        example: Test User
        type: string
      email:
        example: testuser001@example.com
        type: string
      password:
        example: my_SECURE_password123@
        minLength: 8
        type: string
      username:
        example: testuser001
        type: string
    required:
    - display_name
    - email
    - password
    - username
    type: object
  user.createUserResponse:
    properties:
      data:
        $ref: '#/definitions/model.User'
      message:
        type: string
    type: object
  user.loginRequest:
    properties:
      password:
        example: my_SECURE_password123@
        minLength: 8
        type: string
      username:
        example: testuser001
        type: string
    required:
    - password
    - username
    type: object
  user.updateProfileRequest:
    properties:
      display_name:
        example: Updated User 001
        type: string
      email:
        example: updatedtestuser001@example.com
        type: string
    required:
    - display_name
    - email
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation for the Bookmark Management service.
  title: Bookmark Management API
  version: "1.2"
paths:
  /health-check:
    get:
      description: Performs a health check and returns the service status.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/healthcheck.healthCheckResponse'
      summary: Health Check
      tags:
      - health
  /v1/bookmarks:
    get:
      description: Retrieve a list of bookmarks for the authenticated user with pagination
        and sorting
      parameters:
      - default: 1
        description: Page number
        example: 1
        in: query
        name: page
        type: integer
      - default: 5
        description: Number of items per page
        example: 5
        in: query
        name: limit
        type: integer
      - default: -created_at
        description: Sorting criteria(
        example: -created_at,description
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/model.Bookmark'
                type: array
              pagination:
                $ref: '#/definitions/common.Paging'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Message'
      security:
      - Bearer: []
      summary: List bookmarks
      tags:
      - Bookmarks
    post:
      consumes:
      - application/json
      description: Create a new bookmark for the authenticated user
      parameters:
      - description: Create Bookmark Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/bookmark.createBookmarkRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/bookmark.createBookmarkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Message'
      security:
      - Bearer: []
      summary: Create a new bookmark
      tags:
      - Bookmarks
  /v1/bookmarks/{id}:
    delete:
      consumes:
      - application/json
      description: Delete bookmark by ID for authenticated user
      parameters:
      - description: Bookmark ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Message'
      security:
      - Bearer: []
      summary: Delete bookmark by ID
      tags:
      - Bookmarks
    put:
      consumes:
      - application/json
      description: Update bookmark by ID for authenticated user
      parameters:
      - description: Bookmark ID
        in: path
        name: id
        required: true
        type: string
      - description: Bookmark update info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bookmark.updateBookmarkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Message'
      security:
      - BearerAuth: []
      summary: Update bookmark by ID
      tags:
      - Bookmarks
  /v1/generate-password:
    get:
      description: Generates a cryptographically secure random password of fixed length.
      produces:
      - text/plain
      responses:
        "200":
          description: w7h3Q9FeXskn
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Generate a random password
      tags:
      - password
  /v1/links/redirect/{code}:
    get:
      consumes:
      - application/json
      description: Retrieves the original URL from a given shortened code.
      parameters:
      - description: Shortened code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/link.shortenURLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/link.shortenURLResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/link.shortenURLResponse'
      summary: Get original URL
      tags:
      - URL
  /v1/links/shorten:
    post:
      consumes:
      - application/json
      description: Shortens a given URL and returns a unique code.
      parameters:
      - description: URL to shorten
        in: body
        name: shortenURLRequest
        required: true
        schema:
          $ref: '#/definitions/link.shortenURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/link.shortenURLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/link.shortenURLResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/link.shortenURLResponse'
      summary: Shorten a URL
      tags:
      - URL
  /v1/self/info:
    get:
      description: Retrieve the profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.createUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Message'
      security:
      - Bearer: []
      summary: Get user profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update the profile of the authenticated user
      parameters:
      - description: Updated user profile
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/user.updateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Message'
      security:
      - Bearer: []
      summary: Update user profile
      tags:
      - Users
  /v1/users/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/user.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: string
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Message'
      summary: User login
      tags:
      - Users
  /v1/users/register:
    post:
      consumes:
      - application/json
      description: Create a new user with the provided information
      parameters:
      - description: User to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.createUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.createUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Message'
      summary: Create a new user
      tags:
      - Users
schemes:
- http
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"